+++ tests/helpers/pretender.js
@@ -1,5 +1,6 @@
 var stubServer = function() {
   var pretender = new Pretender();
+  DS._routes = Ember.create(null);
 
   pretender.unhandledRequest = function(verb, path, request) {
     var string = "Pretender: non-existing "+verb+" "+path, request
@@ -12,7 +13,7 @@ var stubServer = function() {
 
     availableRequests: {
       'post': [],
-      'put': []
+      'patch': []
     },
 
     get: function(url, response) {
@@ -43,19 +44,19 @@ var stubServer = function() {
       });
     },
 
-    put: function(url, expectedRequest, response) {
+    patch: function(url, expectedRequest, response) {
       var _this = this;
 
-      this.validatePayload(expectedRequest, 'PUT', url);
-      this.validatePayload(response, 'PUT', url);
+      this.validatePayload(expectedRequest, 'PATCH', url);
+      this.validatePayload(response, 'PATCH', url);
 
-      this.availableRequests.put.push({
+      this.availableRequests.patch.push({
         request: expectedRequest,
         response: response
       });
 
-      this.pretender.put(url, function(request){
-        var responseForRequest = _this.responseForRequest('put', request);
+      this.pretender.patch(url, function(request){
+        var responseForRequest = _this.responseForRequest('patch', request);
 
         var string = JSON.stringify(responseForRequest);
         return [200, {"Content-Type": "application/json"}, string]
@@ -72,14 +73,12 @@ var stubServer = function() {
     responseForRequest: function(verb, currentRequest) {
       var respectiveResponse;
       var availableRequests = this.availableRequests[verb];
-
-      currentRequest = currentRequest.requestBody;
+      var actualRequest = JSON.stringify(JSON.parse(currentRequest.requestBody));
 
       for (requests in availableRequests) {
         if (!availableRequests.hasOwnProperty(requests))
           continue;
 
-        var actualRequest = JSON.stringify(JSON.parse(currentRequest));
         var request = JSON.stringify(availableRequests[requests].request);
         var response = JSON.stringify(availableRequests[requests].response);
 
@@ -93,7 +92,7 @@ var stubServer = function() {
         return respectiveResponse;
       } else {
         var error = "No response defined for "+verb+" request";
-        console.error(error, currentRequest);
+        console.error(error, actualRequest);
 
         if (availableRequests.length) {
           console.log("Current defined requests:");
@@ -121,4 +120,5 @@ var stubServer = function() {
 
 var shutdownFakeServer = function(fakeServer) {
   fakeServer.pretender.shutdown();
+  DS._routes = Ember.create(null);
 }