+++ tests/unit/adapter/build-url-test.js
@@ -22,14 +22,3 @@ test("simple replacement", function() {
   DS._routes["comment"] = "posts/comments/{id}";
   equal(adapter.buildURL('comment', 1), '/posts/comments/1');
 });
-
-// TODO, actually support this
-// test("captured link", function() {
-//   DS._routes["comment"] = "posts/{posts.id}/comments";
-//   equal(adapter.buildURL('comment', 1), '/posts/2/comments/1');
-// });
-//
-// test("exploding", function() {
-//   DS._routes["comment"] = "comments/{posts.comments}";
-//   equal(adapter.buildURL('comment', 1), '/posts/2/comments/1');
-// });